+++ src/iop/colorout.c
@@ -393,8 +393,8 @@ process (struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, void *ivoi
     for(int j=0; j<roi_out->height; j++)
     {
 
-      float *in  = (float*)ivoid + ch*roi_in->width *j;
-      float *out = (float*)ovoid + ch*roi_out->width*j;
+      float *in  = (float*)ivoid + (size_t)ch*roi_in->width *j;
+      float *out = (float*)ovoid + (size_t)ch*roi_out->width*j;
       const __m128 m0 = _mm_set_ps(0.0f,d->cmatrix[6],d->cmatrix[3],d->cmatrix[0]);
       const __m128 m1 = _mm_set_ps(0.0f,d->cmatrix[7],d->cmatrix[4],d->cmatrix[1]);
       const __m128 m2 = _mm_set_ps(0.0f,d->cmatrix[8],d->cmatrix[5],d->cmatrix[2]);
@@ -415,8 +415,8 @@ process (struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, void *ivoi
     for(int j=0; j<roi_out->height; j++)
     {
 
-      float *in  = (float*)ivoid + ch*roi_in->width *j;
-      float *out = (float*)ovoid + ch*roi_out->width*j;
+      float *in  = (float*)ivoid + (size_t)ch*roi_in->width *j;
+      float *out = (float*)ovoid + (size_t)ch*roi_out->width*j;
 
       for(int i=0; i<roi_out->width; i++, in+=ch, out+=ch )
       {
@@ -443,7 +443,7 @@ process (struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, void *ivoi
       float Lab[rowsize];
       float rgb[rowsize];
 
-      const int m=(k*(roi_out->width*ch));
+      const size_t m = (size_t)k*roi_out->width*ch;
       for (int l=0; l<roi_out->width; l++)
       {
         int li=3*l,ii=ch*l;
