+++ src/iop/useless_simple.c
@@ -102,8 +102,8 @@ void process(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, void *
 #endif
   for(int j=0; j<roi_out->height; j++)
   {
-    float *in  = ((float *)i) + ch*roi_in->width *j;
-    float *out = ((float *)o) + ch*roi_out->width*j;
+    float *in  = ((float *)i) + (size_t)ch*roi_in->width *j;   // make sure to address input, output and temp buffers with size_t as we want to also
+    float *out = ((float *)o) + (size_t)ch*roi_out->width*j;   // correctly handle huge images 
     for(int i=0; i<roi_out->width; i++)
     {
       // calculate world space coordinates:
